+++ docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # GoNB Changelog
 
+## Next
+
+* Feature request #138
+  * Added openssh-client, rsync and curl, to allow users to install other dependencies.
+  * Added sudo for apt install and apt update.
+  * Added support for `autostart.sh` that if present in the mounted container `/notebooks` directory, and if root owned
+    and set as executable.
+* Updated Dockerfile to latest version to JupyterLab -- now the base docker is served `quay.io/jupyter/base-notebook`
+
 ## v0.10.5, Added SendAsDownload
 
 * Added `dom.SendAsDownload` to send data from cells to the client by triggering a browser download. #134